backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scapy==2.5.0
psutil==5.9.6
openpyxl==3.1.2
python-nmap==0.7.1
netifaces==0.11.0